+++ CHANGELOG.md
@@ -22,6 +22,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - (Breaking) Near point data in distance tracking function callbacks are in incorrect format - was an array instead of an object of `{ x, y }`.
 
+## [0.6.3] - 2022-11-17
+
+Add `desiredDistanceMonitorAllowRecalibrate` option for `.trackDistance()` to allow hiding the recalibration button for the nudger. `.nudgeDistance()` also takes a new second argument, `allowRecalibrate` (default to `true`).
+
 ## [0.6.2] - 2022-06-15
 
 Add `.performanceCompute([callback])` that can run quietly in the background to get `computeArrayFillMHz` and `computeRandomMHz` without showing any Remote Calibrator visual elements.
@@ -358,7 +362,8 @@ No new feature updates in this release. Updated dependency packages and the lice
 
 The framework and some basic functions, e.g., screen size calibration. Released for integration testing.
 
-[unreleased]: https://github.com/EasyEyes/remote-calibrator/compare/v0.6.2...develop
+[unreleased]: https://github.com/EasyEyes/remote-calibrator/compare/v0.6.3...develop
+[0.6.3]: https://github.com/EasyEyes/remote-calibrator/compare/v0.6.2...v0.6.3
 [0.6.2]: https://github.com/EasyEyes/remote-calibrator/compare/v0.6.1...v0.6.2
 [0.6.1]: https://github.com/EasyEyes/remote-calibrator/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/EasyEyes/remote-calibrator/compare/v0.5.1...v0.6.0